+++ gulp/tasks/html2js.js
@@ -89,78 +89,19 @@ module.exports = function( config )
 	} );
 
 	/**
-	 * Compile section component template files.
+	 * Compile section partials.
 	 */
 	config.sections.forEach( function( section )
 	{
-		gulp.task( 'html2js:' + section + ':components', function()
+		gulp.task( 'html2js:' + section + ':partials', function()
 		{
-			// We don't include any section folders that are being built into a separate module.
-			var excludeApp = [];
-			if ( config.modules ) {
-				_.forEach( config.modules, function( moduleDefinition )
-				{
-					if ( moduleDefinition.components ) {
-						moduleDefinition.components.forEach( function( component )
-						{
-							excludeApp.push( '!src/' + section + '/components/' + component + '/**/*.html' );
-						} );
-					}
-				} );
-			}
-
-			return gulp.src( _.union( [
-					'src/' + section + '/components/**/*.html',
-
-					// No form templates.
-					'!src/' + section + '/components/forms/**/*.html',
-
-					// Include partials.
-					'src/' + section + '/views/**/_*.html',
-
-				], excludeApp ), { base: 'src' } )
+			return gulp.src( [ 'src/' + section + '/views/**/_*.html' ], { base: 'src' } )
 				.pipe( plugins.ngHtml2js( {
 					moduleName: 'App',
 					prefix: '/',
 				} ) )
 				.pipe( plugins.rename( { extname: '.html.js' } ) )
-				.pipe( gulp.dest( config.buildDir + '/tmp/' + section + '-component-templates/' ) );
+				.pipe( gulp.dest( config.buildDir + '/tmp/' + section + '-partials/' ) );
 		} );
 	} );
-
-	/**
-	 * Compile module template files.
-	 */
-	var moduleBuilds = [];
-	if ( config.modules ) {
-
-		// We loop through all of the extra modules we need to build and set up gulp tasks to build them.
-		_.forEach( config.modules, function( moduleDefinition, outputFilename )
-		{
-			// Skip if we don't have a component or module.
-			if ( !moduleDefinition.components || !moduleDefinition.module ) {
-				return;
-			}
-
-			// Gotta build all components in the module separately.
-			moduleDefinition.components.forEach( function( component )
-			{
-				var taskName = 'html2js:module:' + outputFilename + ':' + component;
-				moduleBuilds.push( taskName );
-
-				gulp.task( taskName, function()
-				{
-					return gulp.src( [ 'src/app/components/' + component + '/**/*.html' ] )
-						.pipe( plugins.ngHtml2js( {
-							moduleName: moduleDefinition.module,
-							prefix: '/app/components/' + component + '/'
-						} ) )
-						.pipe( plugins.rename( { extname: '.html.js' } ) )
-						.pipe( gulp.dest( config.buildDir + '/tmp/module-templates/' + component ) );
-				} );
-			} );
-		} );
-	}
-
-	gulp.task( 'html2js:modules', moduleBuilds );
 };